// HSSIS.Web.UI/ClientApp/src/app/_enums/common-enums.ts
export enum ResponseMessageTypeEnum {
    Undefined = 0,
    Success = 1,
    Error = 2,
    Warning = 3,
    Fail = 4,
    ValidationFailed = 5,
    Information = 6
  }
  
  export enum ResponseMessageCancelTypeEnum {
    AutoClose = 0,
    ManualClose = 1
  }
  
  export enum ResponseMessageDisplayTypeEnum {
    SnackBarMessage = 0,
    PopupMessage = 1
  }
  